Isolate panic_on_next_alloc tests into per-process binaries
The panic_on_next_alloc flag is a process-global atomic that the tracking
global allocator consumes on the next allocation on any thread. When both
panic-on-alloc tests share one test binary, libtest's multi-threaded harness
keeps other threads live; an allocation on one of them can trip the global
flag while a test holds it enabled. #[serial] only serializes the two test
bodies, so it cannot prevent this: it caused rare hangs and a failed
catch_unwind assertion in CI.

Split the two tests into separate integration test files so each runs as the
only test in its own process, giving true per-process isolation. With one test
per binary the #[serial] guard is unnecessary, so drop it along with the now
unused serial_test dev-dependency.
