+++ packages/alloc_tracker/Cargo.toml
@@ -10,10 +10,6 @@ license.workspace = true
 repository.workspace = true
 rust-version.workspace = true
 
-[package.metadata.cargo-udeps.ignore]
-# Depends on enabled features and dev-dependencies cannot be optional.
-development = ["serial_test"]
-
 [package.metadata.docs.rs]
 all-features = true
 
@@ -29,7 +25,6 @@ serde_json = { workspace = true }
 [dev-dependencies]
 criterion = { workspace = true }
 mutants = { workspace = true }
-serial_test = { workspace = true }
 static_assertions = { workspace = true }
 tempfile = { workspace = true }
